+++ examples/authentication/mui/src/pages/posts/index.tsx
@@ -1,4 +1,3 @@
 export * from "./list";
 export * from "./create";
 export * from "./edit";
-export * from "./show";