+++ packages/mui/src/components/index.tsx
@@ -2,9 +2,10 @@
 export { Layout } from "./layout";
 export { Sider } from "./layout/sider";
 export { Title } from "./layout/title";
+export { Header } from "./layout/header";
 
 // Pages
-export { LoginPage, ReadyPage, ErrorComponent } from "./pages";
+export { LoginPage, ReadyPage, ErrorComponent, AuthPage } from "./pages";
 
 export * from "./crud";
 export * from "./buttons";