+++ packages/mui/src/components/pages/auth/components/index.tsx
@@ -1,0 +1,4 @@
+export { LoginPage } from "./login";
+export { RegisterPage } from "./register";
+export { ForgotPasswordPage } from "./forgotPassword";
+export { UpdatePasswordPage } from "./updatePassword";